Moved cqc download jobs and utils to new folder structure
