+++ projects/_01_ingest/cqc_api/jobs/bulk_download_cqc_providers.py
@@ -1,7 +1,7 @@
 import json
 from datetime import date
 
-from utils import cqc_api as cqc
+from projects._01_ingest.cqc_api.utils import cqc_api as cqc
 from schemas.cqc_provider_schema import PROVIDER_SCHEMA
 from utils import aws_secrets_manager_utilities as ars
 from utils import utils
